# apidoc.yaml
swagger: '2.0'
...
info:
  description: work.muxixyz.com API 
  version: 1.0.0
  title: Work API
  contact:
    email: shiina_orez@qq.com
  #license
### information end ###
host: work.muxixyz.com
basePath: /api/v1.0
schemes:
  - https
  - http

### tag start ###
...
securityDefinitions:

  basic:
    type: basic
    description: http basic auth,即在请求头中添加Authorization字段，内容为base64.encode(“邮箱:muxi304”)

  token:
    type: apiKey
    description: token,然后在请求头中添加Authorization字段，内容为base64.encode(“token:”)
    name: token
    in: header

### auth end ###

### path start ###
paths:
  /auth/signup/:
    post:
      tags:
        - auth
      summary: Signup a account.
      description: Just save a list of user.
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: POSTdata
          in: body
          description: data to fill the segments
          schema:
            required:
              - name
              - email
            properties:
              name:
                type: string
              email:
                type: string
              avatar:
                type: string
              tel:
                type: string
              teamID:
                type: integer
      responses:
        200:
          description: Signup successfully
        401:
          description: User already exist
  /auth/login/:
    post:
      tags:
        - auth
      summary: Login a account
      description: Get a token formed by user-info
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: POSTdata
          in: body
          description: data to mark who want a token
          schema:
            required:
              - username
            properties:
              username:
                type: string
      responses:
        200:
          description: Return a token
          schema:
            required:
              - token
            properties:
              token:
                type: string
  /auth/verify/:
     post:
      tags:
        - auth
      summary: API for verify the token
      description: verify token out of ddl
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: MetaData
          in: body
          description: token
          schema:
            required:
              - token
            properties:
              token:
                type: string
      responses:
        200:
          description: OK
          schema:
            required:
              - uid
            properties:
              uid:
                type: integer
        401:
          description: FAIL
...